Refactor: Amount 타입을 double에서 int64_t로 변경
한국 원화는 소수점이 없으므로 정수 타입 사용이 더 적합합니다.
이를 통해 부동소수점 오차를 제거하고 타입 안전성을 확보했습니다.

- DTO/Entity: Amount를 int64_t로 변경
- SQLite 스키마: REAL → INTEGER 타입 변경
- Storage: bind_double → bind_int64, column_double → column_int64
- View: 포맷 지정자 %ld → %lld, ToDouble → ToLongLong
- 캐스팅 제거: int64_t 네이티브 타입 사용으로 명시적 캐스팅 불필요

주의: 기존 .ml 파일과 호환되지 않음 (스키마 변경)

+++ MissionLedger/src/MLDefine.h
@@ -1,5 +1,6 @@
 ﻿#pragma once
 #include <string>
+#include <cstdint>
 
 enum class E_MLTransactionType
 {
@@ -13,7 +14,7 @@ struct FMLTransactionData {
     std::string Category;
     std::string Item;
     std::string Description;
-    double Amount;
+    int64_t Amount;
     std::string ReceiptNumber;
     std::string DateTime;  // 입력 시 빈 문자열, 출력 시 포맷된 날짜/시간 문자열
 };
